Refactoring the code
- move export function/code to `src/lib`
- only store client/html related code in `src/js`
- can be used as npm module now
