+++ package.json
@@ -2,7 +2,7 @@
   "name": "ds2-tfjs",
   "version": "1.0.0",
   "description": "",
-  "main": "dist/index.js",
+  "main": "dist/lib/index.js",
   "engines": {
     "node": "8.*"
   },
@@ -27,10 +27,10 @@
   },
   "scripts": {
     "build": "tsc && rollup -c",
-    "start": "node dist/index.js",
+    "start": "node dist/server/index.js",
     "watch:server": "tsc -w",
     "watch:js": "rollup --config rollup.config.js --watch",
-    "dev": "concurrently --names \"serverjs,clientjs,server\" -c \"bgBlue.bold,bgMagenta.bold\" npm:watch:server npm:watch:js \"nodemon dist/index.js\"",
+    "dev": "concurrently --names \"serverjs,clientjs,server\" -c \"bgBlue.bold,bgMagenta.bold\" npm:watch:server npm:watch:js \"nodemon dist/server/index.js\"",
     "lint": "tslint -p tslint-tsconfig.json -c tslint.json src/**/*.ts"
   },
   "repository": {
