+++ src/lib/constants.ts
@@ -1,4 +1,6 @@
+// Blank index (-) in the CHAR_MAP
 export const BLANK_INDEX = 28;
+// Character list
 export const CHAR_MAP = ' abcdefghijklmnopqrstuvwxyz\'-';
 export const EPSILON = 1e-5;
 export const DEFAULT_TRIE_WEIGHT = 0.2;
