+++ src/lib/index.ts
@@ -1,0 +1,3 @@
+export * from './webaudiol16';
+export * from './constants';
+export * from './deepspeech2';
