+++ src/lib/trie_pb.d.ts
